Health Bar Update
The tool bar now has health bars for the player and the world.
I added an escape key icon and text to inform the player they can go
back to the title screen with escape. This will probably open a menu
later.
Changed health in player object to a float to make it work better with
the health bar.